+++ configure
@@ -1,0 +1,10 @@
+#!/bin/bash
+set -e # to quit on errors
+# This script will install the fido compiler and library on a Unix based system.
+cd src/
+
+if [[ "$1" == "-noSFML" ]]; then
+	rm -r FidoKin
+	rm -r FidoSim
+	rm ../Tests/simulator.cpp
+fi
