+++ includes/LSInterpolator.h
