+++ includes/Trainer.h
