+++ src/FidoKin/Emitter.cpp
