+++ src/FidoKin/Emitter.h
